+++ dev/breeze/src/airflow_breeze/global_constants.py
@@ -67,6 +67,11 @@ AUTOCOMPLETE_INTEGRATIONS = sorted(
     ]
 )
 
+# Unlike everything else, k8s versions are supported as long as 2 major cloud providers support them.
+# See:
+#   - https://endoflife.date/amazon-eks
+#   - https://endoflife.date/azure-kubernetes-service
+#   - https://endoflife.date/google-kubernetes-engine
 ALLOWED_KUBERNETES_VERSIONS = ["v1.23.13", "v1.24.7", "v1.25.3", "v1.26.0"]
 ALLOWED_EXECUTORS = ["KubernetesExecutor", "CeleryExecutor", "LocalExecutor", "CeleryKubernetesExecutor"]
 ALLOWED_KIND_OPERATIONS = ["start", "stop", "restart", "status", "deploy", "test", "shell", "k9s"]
@@ -83,7 +88,7 @@ ALLOWED_POSTGRES_VERSIONS = ["11", "12", "13", "14", "15"]
 ALLOWED_MYSQL_VERSIONS = ["5.7", "8"]
 ALLOWED_MSSQL_VERSIONS = ["2017-latest", "2019-latest"]
 
-PIP_VERSION = "23.0"
+PIP_VERSION = "23.0.1"
 
 
 @lru_cache(maxsize=None)